+++ unit-tests/json-to-iac/power-vs-volumes.test.js
@@ -59,7 +59,7 @@ resource "ibm_pi_volume" "oracle_template_volume_oracle_1_db_1" {
   provider               = ibm.power_vs_dal12
   pi_cloud_instance_id   = ibm_resource_instance.power_vs_workspace_oracle_template.guid
   pi_volume_size         = "90"
-  pi_volume_name         = "\${var.prefix}-oracle_template-oracle-1-db-1"
+  pi_volume_name         = "\${var.prefix}-oracle-template-oracle-1-db-1"
   pi_volume_shareable    = true
   pi_replication_enabled = false
   pi_affinity_policy     = "affinity"
@@ -97,7 +97,7 @@ resource "ibm_pi_volume" "oracle_template_volume_oracle_1_db_1" {
   provider                = ibm.power_vs_dal12
   pi_cloud_instance_id    = ibm_resource_instance.power_vs_workspace_oracle_template.guid
   pi_volume_size          = "90"
-  pi_volume_name          = "\${var.prefix}-oracle_template-oracle-1-db-1"
+  pi_volume_name          = "\${var.prefix}-oracle-template-oracle-1-db-1"
   pi_volume_shareable     = true
   pi_replication_enabled  = false
   pi_affinity_policy      = "anti-affinity"
@@ -134,7 +134,7 @@ resource "ibm_pi_volume" "oracle_template_volume_oracle_1_db_1" {
   provider               = ibm.power_vs_dal12
   pi_cloud_instance_id   = ibm_resource_instance.power_vs_workspace_oracle_template.guid
   pi_volume_size         = "90"
-  pi_volume_name         = "\${var.prefix}-oracle_template-oracle-1-db-1"
+  pi_volume_name         = "\${var.prefix}-oracle-template-oracle-1-db-1"
   pi_volume_shareable    = true
   pi_replication_enabled = false
   pi_storage_pool        = "Tier1-Flash-4"